requirements.txt
streamlit
joblib
scikit-learn
pandas
numpy
xgboost
lightgbm
pydeck
psycopg2-binary
fastapi
uvicorn
